Split out task item component
